split file embeddings, fix the relevant files
